pyproject.toml
[project]
name = "tool-da"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.27.0",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "duckduckgo-search>=8.1.1",
    "feedparser>=6.0.12",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
]
